+++ src/components/UiButton/UiButton.js
@@ -1,57 +1,10 @@
-import { trimSlotText, propValidator } from '@/utils/helpers';
+import { isBoolean, isString } from '@/utils/inspect';
+import { getSlot, trimSlotText, asyncDelay } from '@/utils/helpers';
 import routable from '@/mixins/routable';
-import { factoryStylable } from '@/mixins/stylable';
-import defaultStyles from './UiButton.scss';
-import themePrimaryStyles from './UiButton.themePrimary.scss';
-import themeSecondaryStyles from './UiButton.themeSecondary.scss';
-import themeSuccessStyles from './UiButton.themeSuccess.scss';
-import themeDangerStyles from './UiButton.themeDanger.scss';
-import themeWarningStyles from './UiButton.themeWarning.scss';
-import themeInfoStyles from './UiButton.themeInfo.scss';
-import themeLightStyles from './UiButton.themeLight.scss';
-import themeDarkStyles from './UiButton.themeDark.scss';
-import themeSilentStyles from './UiButton.themeSilent.scss';
-
-const themesStyles = {
-  primary: themePrimaryStyles,
-  secondary: themeSecondaryStyles,
-  success: themeSuccessStyles,
-  danger: themeDangerStyles,
-  warning: themeWarningStyles,
-  info: themeInfoStyles,
-  light: themeLightStyles,
-  dark: themeDarkStyles,
-  silent: themeSilentStyles,
-};
-const defaultTheme = 'secondary';
-
-const tagValidator = propValidator('tag', [
-  'button',
-  'a',
-  'div',
-]);
-const typeValidator = propValidator('type', [
-  'button',
-  'submit',
-  'reset',
-]);
-const sizeValidator = propValidator('size', [
-  'small',
-  'medium',
-  'large',
-]);
-const contentAlignValidator = propValidator('contentAlign', [
-  'left',
-  'center',
-  'right',
-  'justify',
-]);
-
-const stylable = factoryStylable({
-  defaultStyles,
-  themesStyles,
-  defaultTheme,
-});
+import {
+  stylable,
+  props,
+} from './options';
 
 export default {
   name: 'UiButton',
@@ -59,74 +12,27 @@ export default {
     routable,
     stylable,
   ],
-  props: {
-    tag: {
-      type: String,
-      default: 'button',
-      ...tagValidator,
-    },
-    type: {
-      type: String,
-      default: 'button',
-      ...typeValidator,
-    },
-    disabled: {
-      type: Boolean,
-      default: false,
-    },
-    size: {
-      type: String,
-      default: 'medium',
-      ...sizeValidator,
-    },
-    fluid: {
-      type: Boolean,
-      default: false,
-    },
-    round: {
-      type: Boolean,
-      default: false,
-    },
-    circle: {
-      type: Boolean,
-      default: false,
-    },
-    squared: {
-      type: Boolean,
-      default: false,
-    },
-    hasIcon: {
-      type: Boolean,
-      default: false,
-    },
-    active: {
-      type: Boolean,
-      default: false,
-    },
-    contentAlign: {
-      type: String,
-      default: 'center',
-      ...contentAlignValidator,
-    },
-    empty: {
-      type: Boolean,
-      default: false,
-    },
-  },
+  props,
 
   computed: {
     attributes() {
-      if (this.tag !== 'button') return null;
-      return { type: this.type };
+      const type = (this.tag === 'button')
+        ? { type: this.type }
+        : null;
+
+      return {
+        ...type,
+      };
     },
     classesRoot() {
-      const { size, contentAlign } = this;
+      const { size, square, contentAlign } = this;
       return {
         [this.styles.root]: true,
         [this.styles.isFluid]: this.fluid,
         [this.styles.isRound]: this.round,
         [this.styles.isCircle]: this.circle,
-        [this.styles.isSquared]: this.squared,
+        [this.styles.isSquare]: isBoolean(square) && square,
+        [this.styles[`square_${square}`]]: isString(square),
         [this.styles.isActive]: this.active,
         [this.styles.hasIcon]: this.hasIcon,
         [this.styles.isDisabled]: this.disabled,
@@ -138,31 +44,51 @@ export default {
   },
 
   methods: {
+    async handleClick() {
+      this.click();
+      await asyncDelay(300);
+      this.$el.blur();
+    },
+
     genRoot(childNodes = []) {
       const { tag, data } = this.generateRouterLink(this.classesRoot);
+      const listenersKey = this.to ? 'nativeOn' : 'on';
 
-      data.attrs = { ...data.attrs, ...this.attributes };
-
-      return this.$createElement(tag, data, childNodes);
+      return this.$createElement(tag, {
+        ...data,
+        attrs: {
+          ...data.attrs,
+          ...this.attributes,
+        },
+        [listenersKey]: {
+          ...data[listenersKey],
+          click: this.handleClick,
+        },
+      }, childNodes);
     },
+
     genInner(childNodes = []) {
       return this.$createElement('span', {
         class: this.styles.inner,
       }, childNodes);
     },
-    genDefaultSlot() {
-      const defaultSlot = this.$scopedSlots.default;
 
-      if (!defaultSlot) return null;
-
-      return trimSlotText(defaultSlot());
+    genLink(childNodes = []) {
+      return this.$createElement('span', {
+        class: this.styles.link,
+      }, childNodes);
     },
   },
 
   render() {
+    const defaultSlot = trimSlotText(getSlot(this));
+    const content = (this.theme === 'link')
+      ? this.genLink(defaultSlot)
+      : defaultSlot;
+
     return this.genRoot([
       this.genInner([
-        this.genDefaultSlot(),
+        content,
       ]),
     ]);
   },