+++ src/components/UiButtonGroup/index.js
@@ -1,0 +1,4 @@
+import factoryComponentPlugin from '@/utils/factoryComponentPlugin';
+import UiButtonGroup from './UiButtonGroup';
+
+export default factoryComponentPlugin(UiButtonGroup);