+++ src/components/UiCheckbox/index.js
@@ -1,0 +1,4 @@
+import factoryComponentPlugin from '@/utils/factoryComponentPlugin';
+import UiCheckbox from './UiCheckbox';
+
+export default factoryComponentPlugin(UiCheckbox);