+++ src/utils/helpers/propValidator.js
@@ -1,4 +1,4 @@
-import { Logger } from '@/utils/logger';
+import { Logger } from '../logger';
 
 const logger = new Logger({
   scope: 'propValidator',