+++ src/utils/helpers/randomArrayElement.js
@@ -1,5 +1,5 @@
-import { InvalidTypeError } from '@/utils/errors';
-import { isArray } from '@/utils/inspect';
+import { InvalidTypeError } from '../errors';
+import { isArray } from '../inspect';
 import randomInteger from './randomInteger';
 
 // const random = (min, max) => {